pyproject.toml
[tool.poetry]
name = "reddit-pushbullet"
version = "0.1.0"
description = "Initializes the RedditScanner with subreddit, regex, parameters, and flairs, then continuously scans for new posts every 10 minutes."
authors = ["SomaRe"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
praw = "^7.8.1"
pushbullet-python = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
